refactor(Application/Baseball/Util): create beginGame() to use recursive
modify beginGame() to use recursive rather than while loop for better readability

BREAKING CHANGE: Before:

used while loop for running game

After:

use recursive calling for running game
